DOC: Fix the GitHub pages deployment instructions
Fix the GitHub pages deployment instructions: the deployment is done
from the `main` branch, which will take care of deploying it to the
branch configured in GitHub (e.g. `gh-pages`) for the deployment. So
remove the `gh-pages` checkout part.

Take advantage of the commit to fix grammar in the deployment section.